requirements.txt
django==5.0.2
channels==3.0.4
redis==5.0.1
channels-redis==4.2.0
python-dotenv==1.0.1
dj-database-url==2.1.0
django-cors-headers==4.3.1
gunicorn==21.2.0
psycopg2==2.9.9
daphne==3.0.2